+++ packages/generator-widget/CHANGELOG.md
@@ -13,6 +13,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [9.3.0] - 2022-09-22
 
+### Changed
+
 -   We updated templates to use Cypress instead of Wdio for e2e testing purposes.
 -   We updated templates to provide the latest available functionalities.